decouple wizard from cli-utils
