+++ toolkit/partner-chains-cli/src/register/register2.rs
@@ -1,11 +1,9 @@
 use super::RegisterValidatorMessage;
+use super::{PlainPublicKeyParam, SidechainPublicKeyParam, StakePoolSigningKeyParam};
 use crate::cardano_key::get_mc_staking_signing_key_from_file;
 use crate::io::IOContext;
 use crate::CmdRun;
 use clap::Parser;
-use cli_utils::key_params::{
-	PlainPublicKeyParam, SidechainPublicKeyParam, StakePoolSigningKeyParam,
-};
 use sidechain_domain::crypto::cardano_spo_public_key_and_signature;
 use sidechain_domain::UtxoId;
